Additional dead code cleanup.
