+++ src/common/memory_patcher.cpp
@@ -10,25 +10,16 @@
 #include <QJsonArray>
 #include <QJsonDocument>
 #include <QJsonObject>
-#include <QListView>
 #include <QMessageBox>
 #include <QString>
 #include <QXmlStreamReader>
 #include <pugixml.hpp>
 #include "common/logging/log.h"
 #include "common/path_util.h"
-#include "common/singleton.h"
 #include "core/file_format/psf.h"
 #include "memory_patcher.h"
 
 namespace MemoryPatcher {
-
-EXPORT uintptr_t g_eboot_address;
-uint64_t g_eboot_image_size;
-std::string g_game_serial;
-std::string patchFile;
-bool patches_applied = false;
-std::vector<patchInfo> pending_patches;
 
 std::string toHex(u64 value, size_t byteSize) {
     std::stringstream ss;
@@ -118,93 +109,9 @@ std::string convertValueToHex(const std::string type, const std::string valueStr
     return result;
 }
 
-void ApplyPendingPatches();
+std::vector<PendingPatch> readPatches(std::string gameSerial, std::string appVersion) {
+    std::vector<PendingPatch> pending;
 
-void OnGameLoaded() {
-    if (!patchFile.empty()) {
-        std::filesystem::path patchDir = Common::FS::GetUserPath(Common::FS::PathType::PatchesDir);
-
-        auto filePath = (patchDir / patchFile).native();
-
-        pugi::xml_document doc;
-        pugi::xml_parse_result result = doc.load_file(filePath.c_str());
-
-        auto* param_sfo = Common::Singleton<PSF>::Instance();
-        auto app_version = param_sfo->GetString("APP_VER").value_or("Unknown version");
-
-        if (result) {
-            auto patchXML = doc.child("Patch");
-            for (pugi::xml_node_iterator it = patchXML.children().begin();
-                 it != patchXML.children().end(); ++it) {
-
-                if (std::string(it->name()) == "Metadata") {
-                    if (std::string(it->attribute("isEnabled").value()) == "true") {
-                        std::string currentPatchName = it->attribute("Name").value();
-                        std::string metadataAppVer = it->attribute("AppVer").value();
-                        bool versionMatches = metadataAppVer == app_version;
-
-                        auto patchList = it->first_child();
-                        for (pugi::xml_node_iterator patchLineIt = patchList.children().begin();
-                             patchLineIt != patchList.children().end(); ++patchLineIt) {
-
-                            std::string type = patchLineIt->attribute("Type").value();
-                            if (!versionMatches && type != "mask" && type != "mask_jump32")
-                                continue;
-
-                            std::string currentPatchName = it->attribute("Name").value();
-
-                            for (pugi::xml_node_iterator patchLineIt = patchList.children().begin();
-                                 patchLineIt != patchList.children().end(); ++patchLineIt) {
-
-                                std::string type = patchLineIt->attribute("Type").value();
-                                std::string address = patchLineIt->attribute("Address").value();
-                                std::string patchValue = patchLineIt->attribute("Value").value();
-                                std::string maskOffsetStr =
-                                    patchLineIt->attribute("Offset").value();
-                                std::string targetStr = "";
-                                std::string sizeStr = "";
-                                if (type == "mask_jump32") {
-                                    targetStr = patchLineIt->attribute("Target").value();
-                                    sizeStr = patchLineIt->attribute("Size").value();
-                                } else {
-                                    patchValue = convertValueToHex(type, patchValue);
-                                }
-
-                                bool littleEndian = false;
-
-                                if (type == "bytes16" || type == "bytes32" || type == "bytes64") {
-                                    littleEndian = true;
-                                }
-
-                                MemoryPatcher::PatchMask patchMask = MemoryPatcher::PatchMask::None;
-                                int maskOffsetValue = 0;
-
-                                if (type == "mask")
-                                    patchMask = MemoryPatcher::PatchMask::Mask;
-
-                                if (type == "mask_jump32")
-                                    patchMask = MemoryPatcher::PatchMask::Mask_Jump32;
-
-                                if ((type == "mask" || type == "mask_jump32") &&
-                                    !maskOffsetStr.empty()) {
-                                    maskOffsetValue = std::stoi(maskOffsetStr, 0, 10);
-                                }
-
-                                MemoryPatcher::PatchMemory(currentPatchName, address, patchValue,
-                                                           targetStr, sizeStr, false, littleEndian,
-                                                           patchMask, maskOffsetValue);
-                            }
-                        }
-                    }
-                }
-            }
-        } else {
-            LOG_ERROR(Loader, "couldnt patch parse xml : {}", result.description());
-        }
-    }
-    ApplyPendingPatches();
-
-    // We use the QT headers for the xml and json parsing, this define is only true on QT builds
     QString patchDir;
     Common::FS::PathToQString(patchDir, Common::FS::GetUserPath(Common::FS::PathType::PatchesDir));
     QDir dir(patchDir);
@@ -219,20 +126,18 @@ void OnGameLoaded() {
                       folder.toStdString());
             continue;
         }
-
-        QByteArray jsonData = jsonFile.readAll();
+        const QByteArray jsonData = jsonFile.readAll();
         jsonFile.close();
 
-        QJsonDocument jsonDoc = QJsonDocument::fromJson(jsonData);
-        QJsonObject jsonObject = jsonDoc.object();
+        const QJsonDocument jsonDoc = QJsonDocument::fromJson(jsonData);
+        const QJsonObject jsonObject = jsonDoc.object();
 
         QString selectedFileName;
-        QString serial = QString::fromStdString(g_game_serial);
+        const QString serial = QString::fromStdString(gameSerial);
 
         for (auto it = jsonObject.constBegin(); it != jsonObject.constEnd(); ++it) {
-            QString filePath = it.key();
-            QJsonArray idsArray = it.value().toArray();
-
+            const QString filePath = it.key();
+            const QJsonArray idsArray = it.value().toArray();
             if (idsArray.contains(QJsonValue(serial))) {
                 selectedFileName = filePath;
                 break;
@@ -245,321 +150,105 @@ void OnGameLoaded() {
             continue;
         }
 
-        QString filePath = patchDir + "/" + folder + "/" + selectedFileName;
+        const QString filePath = patchDir + "/" + folder + "/" + selectedFileName;
         QFile file(filePath);
         if (!file.open(QIODevice::ReadOnly | QIODevice::Text)) {
             LOG_ERROR(Loader, "Unable to open the file for reading.");
             continue;
         }
-
-        QByteArray xmlData = file.readAll();
+        const QByteArray xmlData = file.readAll();
         file.close();
-
-        QString newXmlData;
 
         QXmlStreamReader xmlReader(xmlData);
 
         bool isEnabled = false;
         std::string currentPatchName;
 
-        auto* param_sfo = Common::Singleton<PSF>::Instance();
-        auto app_version = param_sfo->GetString("APP_VER").value_or("Unknown version");
         bool versionMatches = true;
 
         while (!xmlReader.atEnd()) {
             xmlReader.readNext();
 
-            if (xmlReader.isStartElement()) {
-                QJsonArray patchLines;
-                if (xmlReader.name() == QStringLiteral("Metadata")) {
-                    QString name = xmlReader.attributes().value("Name").toString();
-                    currentPatchName = name.toStdString();
+            if (!xmlReader.isStartElement()) {
+                continue;
+            }
 
-                    QString appVer = xmlReader.attributes().value("AppVer").toString();
+            if (xmlReader.name() == QStringLiteral("Metadata")) {
+                QString name = xmlReader.attributes().value("Name").toString();
+                currentPatchName = name.toStdString();
 
-                    // Check and update the isEnabled attribute
-                    isEnabled = false;
-                    for (const QXmlStreamAttribute& attr : xmlReader.attributes()) {
-                        if (attr.name() == QStringLiteral("isEnabled")) {
-                            isEnabled = (attr.value().toString() == "true");
-                        }
+                const QString appVer = xmlReader.attributes().value("AppVer").toString();
+
+                isEnabled = false;
+                for (const QXmlStreamAttribute& attr : xmlReader.attributes()) {
+                    if (attr.name() == QStringLiteral("isEnabled")) {
+                        isEnabled = (attr.value().toString() == "true");
                     }
-                    versionMatches = (appVer.toStdString() == app_version);
+                }
+                versionMatches = (appVer.toStdString() == appVersion);
+            } else if (xmlReader.name() == QStringLiteral("PatchList")) {
+                while (!xmlReader.atEnd() &&
+                       !(xmlReader.tokenType() == QXmlStreamReader::EndElement &&
+                         xmlReader.name() == QStringLiteral("PatchList"))) {
 
-                } else if (xmlReader.name() == QStringLiteral("PatchList")) {
-                    QJsonArray linesArray;
-                    while (!xmlReader.atEnd() &&
-                           !(xmlReader.tokenType() == QXmlStreamReader::EndElement &&
-                             xmlReader.name() == QStringLiteral("PatchList"))) {
-                        xmlReader.readNext();
-                        if (xmlReader.tokenType() == QXmlStreamReader::StartElement &&
-                            xmlReader.name() == QStringLiteral("Line")) {
-                            QXmlStreamAttributes attributes = xmlReader.attributes();
-                            QJsonObject lineObject;
-                            lineObject["Type"] = attributes.value("Type").toString();
-                            lineObject["Address"] = attributes.value("Address").toString();
-                            lineObject["Value"] = attributes.value("Value").toString();
-                            lineObject["Offset"] = attributes.value("Offset").toString();
-                            if (lineObject["Type"].toString() == "mask_jump32") {
-                                lineObject["Target"] = attributes.value("Target").toString();
-                                lineObject["Size"] = attributes.value("Size").toString();
-                            }
-                            linesArray.append(lineObject);
-                        }
+                    xmlReader.readNext();
+
+                    if (xmlReader.tokenType() != QXmlStreamReader::StartElement ||
+                        xmlReader.name() != QStringLiteral("Line")) {
+                        continue;
                     }
 
-                    patchLines = linesArray;
-                    if (isEnabled) {
-                        foreach (const QJsonValue& value, patchLines) {
-                            QJsonObject lineObject = value.toObject();
-                            QString type = lineObject["Type"].toString();
+                    const QXmlStreamAttributes a = xmlReader.attributes();
+                    const QString type = a.value("Type").toString();
+                    const QString addr = a.value("Address").toString();
+                    QString val = a.value("Value").toString();
+                    const QString offStr = a.value("Offset").toString();
+                    const QString tgt =
+                        (type == "mask_jump32") ? a.value("Target").toString() : QString{};
+                    const QString sz =
+                        (type == "mask_jump32") ? a.value("Size").toString() : QString{};
 
-                            if ((type != "mask" && type != "mask_jump32") && !versionMatches)
-                                continue;
-
-                            QString address = lineObject["Address"].toString();
-                            QString patchValue = lineObject["Value"].toString();
-                            QString maskOffsetStr = lineObject["Offset"].toString();
-
-                            QString targetStr;
-                            QString sizeStr;
-
-                            if (type == "mask_jump32") {
-                                targetStr = lineObject["Target"].toString();
-                                sizeStr = lineObject["Size"].toString();
-                            } else {
-                                patchValue = QString::fromStdString(convertValueToHex(
-                                    type.toStdString(), patchValue.toStdString()));
-                            }
-
-                            bool littleEndian = false;
-
-                            if (type == "bytes16" || type == "bytes32" || type == "bytes64")
-                                littleEndian = true;
-
-                            MemoryPatcher::PatchMask patchMask = MemoryPatcher::PatchMask::None;
-                            int maskOffsetValue = 0;
-
-                            if (type == "mask")
-                                patchMask = MemoryPatcher::PatchMask::Mask;
-
-                            if (type == "mask_jump32")
-                                patchMask = MemoryPatcher::PatchMask::Mask_Jump32;
-
-                            if ((type == "mask" || type == "mask_jump32") &&
-                                !maskOffsetStr.toStdString().empty()) {
-                                maskOffsetValue = std::stoi(maskOffsetStr.toStdString(), 0, 10);
-                            }
-
-                            MemoryPatcher::PatchMemory(
-                                currentPatchName, address.toStdString(), patchValue.toStdString(),
-                                targetStr.toStdString(), sizeStr.toStdString(), false, littleEndian,
-                                patchMask, maskOffsetValue);
-                        }
+                    if (!isEnabled) {
+                        continue;
                     }
+                    if ((type != "mask" && type != "mask_jump32") && !versionMatches) {
+                        continue;
+                    }
+
+                    PendingPatch pp;
+                    pp.modName = currentPatchName;
+                    pp.address = addr.toStdString();
+
+                    if (type == "mask" || type == "mask_jump32") {
+                        if (!offStr.toStdString().empty()) {
+                            pp.maskOffset = std::stoi(offStr.toStdString(), nullptr, 10);
+                        }
+                        pp.mask = (type == "mask") ? MemoryPatcher::PatchMask::Mask
+                                                   : MemoryPatcher::PatchMask::Mask_Jump32;
+                        pp.value = val.toStdString();
+                        pp.target = tgt.toStdString();
+                        pp.size = sz.toStdString();
+                    } else {
+                        pp.value = convertValueToHex(type.toStdString(), val.toStdString());
+                        pp.target.clear();
+                        pp.size.clear();
+                        pp.mask = MemoryPatcher::PatchMask::None;
+                        pp.maskOffset = 0;
+                    }
+
+                    pp.littleEndian = (type == "bytes16" || type == "bytes32" || type == "bytes64");
+                    pending.emplace_back(std::move(pp));
                 }
             }
         }
 
         if (xmlReader.hasError()) {
-            LOG_ERROR(Loader, "Failed to parse XML for {}", g_game_serial);
+            LOG_ERROR(Loader, "Failed to parse XML for {}", gameSerial);
         } else {
-            LOG_INFO(Loader, "Patches loaded successfully, repository: {}", folder.toStdString());
-        }
-        ApplyPendingPatches();
-    }
-}
-
-void AddPatchToQueue(patchInfo patchToAdd) {
-    if (patches_applied) {
-        PatchMemory(patchToAdd.modNameStr, patchToAdd.offsetStr, patchToAdd.valueStr,
-                    patchToAdd.targetStr, patchToAdd.sizeStr, patchToAdd.isOffset,
-                    patchToAdd.littleEndian, patchToAdd.patchMask, patchToAdd.maskOffset);
-        return;
-    }
-    pending_patches.push_back(patchToAdd);
-}
-
-void ApplyPendingPatches() {
-    patches_applied = true;
-    for (size_t i = 0; i < pending_patches.size(); ++i) {
-        const patchInfo& currentPatch = pending_patches[i];
-
-        if (currentPatch.gameSerial != "*" && currentPatch.gameSerial != g_game_serial)
-            continue;
-
-        PatchMemory(currentPatch.modNameStr, currentPatch.offsetStr, currentPatch.valueStr,
-                    currentPatch.targetStr, currentPatch.sizeStr, currentPatch.isOffset,
-                    currentPatch.littleEndian, currentPatch.patchMask, currentPatch.maskOffset);
-    }
-
-    pending_patches.clear();
-}
-
-void PatchMemory(std::string modNameStr, std::string offsetStr, std::string valueStr,
-                 std::string targetStr, std::string sizeStr, bool isOffset, bool littleEndian,
-                 PatchMask patchMask, int maskOffset) {
-    // Send a request to modify the process memory.
-    void* cheatAddress = nullptr;
-
-    if (patchMask == PatchMask::None) {
-        if (isOffset) {
-            cheatAddress = reinterpret_cast<void*>(g_eboot_address + std::stoi(offsetStr, 0, 16));
-        } else {
-            cheatAddress =
-                reinterpret_cast<void*>(g_eboot_address + (std::stoi(offsetStr, 0, 16) - 0x400000));
+            LOG_INFO(Loader, "Patches parsed successfully, repository: {}", folder.toStdString());
         }
     }
-
-    if (patchMask == PatchMask::Mask) {
-        cheatAddress = reinterpret_cast<void*>(PatternScan(offsetStr) + maskOffset);
-    }
-
-    if (patchMask == PatchMask::Mask_Jump32) {
-        int jumpSize = std::stoi(sizeStr);
-
-        constexpr int MAX_PATTERN_LENGTH = 256;
-        if (jumpSize < 5) {
-            LOG_ERROR(Loader, "Jump size must be at least 5 bytes");
-            return;
-        }
-        if (jumpSize > MAX_PATTERN_LENGTH) {
-            LOG_ERROR(Loader, "Jump size must be no more than {} bytes.", MAX_PATTERN_LENGTH);
-            return;
-        }
-
-        // Find the base address using "Address"
-        uintptr_t baseAddress = PatternScan(offsetStr);
-        if (baseAddress == 0) {
-            LOG_ERROR(Loader, "PatternScan failed for mask_jump32 with pattern: {}", offsetStr);
-            return;
-        }
-        uintptr_t patchAddress = baseAddress + maskOffset;
-
-        // Fills the original region (jumpSize bytes) with NOPs
-        std::vector<u8> nopBytes(jumpSize, 0x90);
-        std::memcpy(reinterpret_cast<void*>(patchAddress), nopBytes.data(), nopBytes.size());
-
-        // Use "Target" to locate the start of the code cave
-        uintptr_t jump_target = PatternScan(targetStr);
-        if (jump_target == 0) {
-            LOG_ERROR(Loader, "PatternScan failed to Target with pattern: {}", targetStr);
-            return;
-        }
-
-        // Converts the Value attribute to a byte array (payload)
-        std::vector<u8> payload;
-        for (size_t i = 0; i < valueStr.length(); i += 2) {
-
-            std::string tempStr = valueStr.substr(i, 2);
-            const char* byteStr = tempStr.c_str();
-            char* endPtr;
-            unsigned int byteVal = std::strtoul(byteStr, &endPtr, 16);
-
-            if (endPtr != byteStr + 2) {
-                LOG_ERROR(Loader, "Invalid byte in Value: {}", valueStr.substr(i, 2));
-                return;
-            }
-            payload.push_back(static_cast<u8>(byteVal));
-        }
-
-        // Calculates the end of the code cave (where the return jump will be inserted)
-        uintptr_t code_cave_end = jump_target + payload.size();
-
-        // Write the payload to the code cave, from jump_target
-        std::memcpy(reinterpret_cast<void*>(jump_target), payload.data(), payload.size());
-
-        // Inserts the initial jump in the original region to divert to the code cave
-        u8 jumpInstruction[5];
-        jumpInstruction[0] = 0xE9;
-        s32 relJump = static_cast<s32>(jump_target - patchAddress - 5);
-        std::memcpy(&jumpInstruction[1], &relJump, sizeof(relJump));
-        std::memcpy(reinterpret_cast<void*>(patchAddress), jumpInstruction,
-                    sizeof(jumpInstruction));
-
-        // Inserts jump back at the end of the code cave to resume execution after patching
-        u8 jumpBack[5];
-        jumpBack[0] = 0xE9;
-        // Calculates the relative offset to return to the instruction immediately following the
-        // overwritten region
-        s32 target_return = static_cast<s32>((patchAddress + jumpSize) - (code_cave_end + 5));
-        std::memcpy(&jumpBack[1], &target_return, sizeof(target_return));
-        std::memcpy(reinterpret_cast<void*>(code_cave_end), jumpBack, sizeof(jumpBack));
-
-        LOG_INFO(Loader,
-                 "Applied Patch mask_jump32: {}, PatchAddress: {:#x}, JumpTarget: {:#x}, "
-                 "CodeCaveEnd: {:#x}, JumpSize: {}",
-                 modNameStr, patchAddress, jump_target, code_cave_end, jumpSize);
-        return;
-    }
-
-    if (cheatAddress == nullptr) {
-        LOG_ERROR(Loader, "Failed to get address for patch {}", modNameStr);
-        return;
-    }
-
-    std::vector<unsigned char> bytePatch;
-
-    for (size_t i = 0; i < valueStr.length(); i += 2) {
-        unsigned char byte =
-            static_cast<unsigned char>(std::strtol(valueStr.substr(i, 2).c_str(), nullptr, 16));
-
-        bytePatch.push_back(byte);
-    }
-
-    if (littleEndian) {
-        std::reverse(bytePatch.begin(), bytePatch.end());
-    }
-
-    std::memcpy(cheatAddress, bytePatch.data(), bytePatch.size());
-
-    LOG_INFO(Loader, "Applied patch: {}, Offset: {}, Value: {}", modNameStr,
-             (uintptr_t)cheatAddress, valueStr);
-}
-
-static std::vector<int32_t> PatternToByte(const std::string& pattern) {
-    std::vector<int32_t> bytes;
-    const char* start = pattern.data();
-    const char* end = start + pattern.size();
-
-    for (const char* current = start; current < end; ++current) {
-        if (*current == '?') {
-            ++current;
-            if (*current == '?')
-                ++current;
-            bytes.push_back(-1);
-        } else {
-            bytes.push_back(strtoul(current, const_cast<char**>(&current), 16));
-        }
-    }
-
-    return bytes;
-}
-
-uintptr_t PatternScan(const std::string& signature) {
-    std::vector<int32_t> patternBytes = PatternToByte(signature);
-    const auto scanBytes = static_cast<uint8_t*>((void*)g_eboot_address);
-
-    const int32_t* sigPtr = patternBytes.data();
-    const size_t sigSize = patternBytes.size();
-
-    uint32_t foundResults = 0;
-    for (uint32_t i = 0; i < g_eboot_image_size - sigSize; ++i) {
-        bool found = true;
-        for (uint32_t j = 0; j < sigSize; ++j) {
-            if (scanBytes[i + j] != sigPtr[j] && sigPtr[j] != -1) {
-                found = false;
-                break;
-            }
-        }
-
-        if (found) {
-            foundResults++;
-            return reinterpret_cast<uintptr_t>(&scanBytes[i]);
-        }
-    }
-
-    return 0;
+    return pending;
 }
 
 } // namespace MemoryPatcher
