--- src/core/memory_patcher.h
@@ -1,25 +1,0 @@
-// SPDX-FileCopyrightText: Copyright 2025 shadPS4 Emulator Project
-// SPDX-License-Identifier: GPL-2.0-or-later
-
-#pragma once
-
-#include <filesystem>
-#include <string>
-
-#include "common/memory_patcher.h"
-
-namespace MemoryPatcher {
-
-struct PendingPatch {
-    std::string modName;
-    std::string address;
-    std::string value;
-    std::string target;
-    std::string size;
-    bool littleEndian = false;
-    PatchMask mask = PatchMask::None;
-    int maskOffset = 0;
-};
-
-std::vector<PendingPatch> readPatches(std::string gameSerial, std::string appVersion);
-} // namespace MemoryPatcher
