--- src/input/input_handler.h
@@ -1,47 +1,0 @@
-// SPDX-FileCopyrightText: Copyright 2024 shadPS4 Emulator Project
-// SPDX-License-Identifier: GPL-2.0-or-later
-
-#pragma once
-
-#include "SDL3/SDL_events.h"
-#include "common/types.h"
-#include "input/controller.h"
-
-// +1 and +2 is taken
-#define SDL_MOUSE_WHEEL_UP SDL_EVENT_MOUSE_WHEEL + 3
-#define SDL_MOUSE_WHEEL_DOWN SDL_EVENT_MOUSE_WHEEL + 4
-#define SDL_MOUSE_WHEEL_LEFT SDL_EVENT_MOUSE_WHEEL + 5
-#define SDL_MOUSE_WHEEL_RIGHT SDL_EVENT_MOUSE_WHEEL + 7
-
-#define SDL_GAMEPAD_BUTTON_TOUCHPAD_LEFT SDL_GAMEPAD_BUTTON_COUNT + 1
-#define SDL_GAMEPAD_BUTTON_TOUCHPAD_CENTER SDL_GAMEPAD_BUTTON_COUNT + 2
-#define SDL_GAMEPAD_BUTTON_TOUCHPAD_RIGHT SDL_GAMEPAD_BUTTON_COUNT + 3
-
-#define SDL_EVENT_TOGGLE_FULLSCREEN SDL_EVENT_USER + 1
-#define SDL_EVENT_TOGGLE_PAUSE SDL_EVENT_USER + 2
-#define SDL_EVENT_CHANGE_CONTROLLER SDL_EVENT_USER + 3
-#define SDL_EVENT_TOGGLE_SIMPLE_FPS SDL_EVENT_USER + 4
-#define SDL_EVENT_RELOAD_INPUTS SDL_EVENT_USER + 5
-#define SDL_EVENT_MOUSE_TO_JOYSTICK SDL_EVENT_USER + 6
-#define SDL_EVENT_MOUSE_TO_GYRO SDL_EVENT_USER + 7
-#define SDL_EVENT_RDOC_CAPTURE SDL_EVENT_USER + 8
-#define SDL_EVENT_QUIT_DIALOG SDL_EVENT_USER + 9
-#define SDL_EVENT_MOUSE_WHEEL_OFF SDL_EVENT_USER + 10
-
-#define LEFTJOYSTICK_HALFMODE 0x00010000
-#define RIGHTJOYSTICK_HALFMODE 0x00020000
-#define BACK_BUTTON 0x00040000
-
-#define KEY_TOGGLE 0x00200000
-#define MOUSE_GYRO_ROLL_MODE 0x00400000
-
-#define HOTKEY_FULLSCREEN 0xf0000001
-#define HOTKEY_PAUSE 0xf0000002
-#define HOTKEY_SIMPLE_FPS 0xf0000003
-#define HOTKEY_QUIT 0xf0000004
-#define HOTKEY_RELOAD_INPUTS 0xf0000005
-#define HOTKEY_TOGGLE_MOUSE_TO_JOYSTICK 0xf0000006
-#define HOTKEY_TOGGLE_MOUSE_TO_GYRO 0xf0000007
-#define HOTKEY_RENDERDOC 0xf0000008
-
-#define SDL_UNMAPPED UINT32_MAX - 1
