+++ src/qt_gui/main_window.cpp
@@ -15,13 +15,12 @@
 #include "check_update.h"
 #endif
 #include "common/logging/log.h"
+#include "common/memory_patcher.h"
 #include "common/path_util.h"
 #include "common/scm_rev.h"
 #include "control_settings.h"
-#include "core/memory_patcher.h"
 #include "game_install_dialog.h"
 #include "hotkeys.h"
-#include "input/input_handler.h"
 #include "ipc/ipc_client.h"
 #include "kbm_gui.h"
 #include "main_window.h"
